refact: organize integration test in a directory
